+++ nix/lib/external.nix
@@ -12,6 +12,4 @@
       }) { inherit lib; },
     }:
     infuse.v1.infuse;
-
-  conanProfileShow = nixpkgs: nixpkgs.callPackage ../packages/conan/package.nix { };
 }